refactor: extract core to _internal module
