+++ src/conformly/_internal/api.py
@@ -1,15 +1,15 @@
 from typing import Any
 
-from ._internal.generator import (
+from .generator import (
     GenerationContext,
     create_context,
     generate_invalid,
     generate_valid,
 )
-from ._internal.parser import ModelSpec, parse_model
-from ._internal.planner import PlannedTask, plan_violation_task, select_paths
-from ._internal.resolver import ResolvedModel, resolve_model
-from ._internal.types import CasesStrategy, CaseStrategy, ViolationType
+from .parser import ModelSpec, parse_model
+from .planner import PlannedTask, plan_violation_task, select_paths
+from .resolver import ResolvedModel, resolve_model
+from .types import CasesStrategy, CaseStrategy, ViolationType
 
 
 def _ensure_model_or_spec(
